refactor echo plugin
