+++ plugins/echov4/go.mod
@@ -1,14 +1,15 @@
-module github.com/apache/skywalking-go/plugins/echo
+module github.com/apache/skywalking-go/plugins/echov4
 
 go 1.20
 
-require github.com/labstack/echo/v4 v4.11.1
+require (
+	github.com/labstack/echo/v4 v4.11.1
+)
 
 require (
 	github.com/labstack/gommon v0.4.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
-	github.com/stretchr/testify v1.8.2 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 	github.com/valyala/fasttemplate v1.2.2 // indirect
 	golang.org/x/crypto v0.11.0 // indirect
