+++ plugins/echov4/interceptor.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package echo
+package echov4
 
 import (
 	"fmt"
