+++ test/plugins/scenarios/echov4/bin/startup.sh
