+++ tools/go-agent/instrument/plugins/register.go
@@ -20,7 +20,7 @@ package plugins
 import (
 	"github.com/apache/skywalking-go/plugins/core/instrument"
 	"github.com/apache/skywalking-go/plugins/dubbo"
-	"github.com/apache/skywalking-go/plugins/echo"
+	"github.com/apache/skywalking-go/plugins/echov4"
 	"github.com/apache/skywalking-go/plugins/gin"
 	goredisv9 "github.com/apache/skywalking-go/plugins/go-redisv9"
 	"github.com/apache/skywalking-go/plugins/go-restfulv3"
@@ -48,7 +48,7 @@ func init() {
 	registerFramework(goredisv9.NewInstrument())
 	registerFramework(mongo.NewInstrument())
 	registerFramework(runtime_metrics.NewInstrument())
-	registerFramework(echo.NewInstrument())
+	registerFramework(echov4.NewInstrument())
 
 	// gorm related instruments
 	registerFramework(gorm_entry.NewInstrument())
